backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
djangorestframework-camel-case
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
gunicorn
smtplib